Extract authorisation servers logic into separate file
Extract auth server code from ob-directory into new file.

Update references to moved functions.

Refactor list auth server test to put stubbing in a function.
