+++ app/authorisation-servers/index.js
@@ -1,0 +1,9 @@
+const {
+  allAuthorisationServers,
+  authorisationServersForClient,
+  storeAuthorisationServers,
+} = require('./authorisation-servers');
+
+exports.allAuthorisationServers = allAuthorisationServers;
+exports.authorisationServersForClient = authorisationServersForClient;
+exports.storeAuthorisationServers = storeAuthorisationServers;
